refactor(examples): split boundary into a package and a binding layer
`examples/boundary` is now two projects. `Boundary` is plain Julia with no
JLWInterop dependency and no entry points; `lib/` holds the `@api`
declarations, the carrier methods for `Extent`, and the hand-written
`Base.@ccallable` entrypoints, and depends on the package next door.

This is the layout a package gets bindings in without taking on a binding
dependency, and the one someone who does not maintain a package can use to
wrap it. It relies on a declaration defining nothing: `lib/` imports the
wrapped functions by name, which would add methods to them if `@api` took a
body.

Both `[sources]` paths in `lib/Project.toml` are relative, so the file is
committable — `build_library` compiles a temporary copy with them made
absolute. That retires `build.jl`'s `prepare_project`, which hand-rolled the
same rewrite to keep an absolute path out of the tree.

The example smoke test builds `boundary` from `lib/` and `ols` from its own
`src/`, so both layouts stay covered.
