+++ JuliaLibWrapping/examples/boundary/build.jl
@@ -1,53 +1,34 @@
 # Build the boundary example library end-to-end with a bundled Python package.
 # Run from this directory with a recent enough Julia 1.13:
 #
-#   julia --project=. build.jl
+#   julia build.jl
 #
 # Output lands in `out/`. The build-env at `./build-env/` must be
 # instantiated once:
 #
 #   julia --project=build-env -e 'using Pkg; Pkg.instantiate()'
 #
-# In a tutorial-shaped library (post-registration of JuliaLibWrapping and
-# JLWInterop) this script collapses to:
+# The library is built from `lib/`, the binding layer, not from the package
+# beside it. `lib/Project.toml` names both the package and JLWInterop as
+# relative `[sources]` paths, which `build_library` resolves by compiling a
+# temporary copy with those paths made absolute — so nothing here or in that
+# file carries a machine-specific path.
 #
-#     push!(LOAD_PATH, joinpath(@__DIR__, "build-env"))
-#     using JuliaLibWrapping, JuliaC
-#     standard_build(@__DIR__; libname = "boundary", verbose = true)
+# Once JuliaLibWrapping and JLWInterop are registered, the `[sources]` entries
+# go and this script is the four lines below.
 #
 # `using JuliaC` is what activates JuliaLibWrapping's weak dependency
 # on JuliaC.jl — without it, `build_library` errors with a hint.
-#
-# The extra machinery below exists only because we are dogfooding against
-# the in-tree `JLWInterop/` checkout: `prepare_project` adds its path to a
-# temporary copy of `Project.toml`.
-# Once `JLWInterop` is registered, the `prepare_project` step disappears too.
-
-using TOML: TOML
 
 const HERE = @__DIR__
-const JLW_INTEROP = abspath(joinpath(HERE, "..", "..", "..", "JLWInterop"))
-
-function prepare_project()
-    toml = TOML.parsefile(joinpath(HERE, "Project.toml"))
-    sources = get(toml, "sources", Dict{String, Any}())
-    sources["JLWInterop"] = Dict("path" => JLW_INTEROP)
-    toml["sources"] = sources
-    tmp = mktempdir(; prefix = "boundary-project-")
-    open(joinpath(tmp, "Project.toml"), "w") do io
-        TOML.print(io, toml; sorted = true)
-    end
-    cp(joinpath(HERE, "src"), joinpath(tmp, "src"))
-    return tmp
-end
 
 push!(LOAD_PATH, joinpath(HERE, "build-env"))
 using JuliaLibWrapping, JuliaC
 
 result = standard_build(
-    HERE;
+    joinpath(HERE, "lib");
     libname = "boundary",
-    project = prepare_project(),
+    out = joinpath(HERE, "out"),
     verbose = true,
 )
